+++ CHANGES.md
@@ -8,7 +8,8 @@
 
 **breaking change**
 
-* renamed `OptionalHeaders`, `MimeTypes` and `ImageDrivers` to the singular form.
+* renamed `OptionalHeaders`, `MimeTypes` and `ImageDrivers` enums to the singular form. (https://github.com/developmentseed/titiler/pull/258)
+* renamed `MimeType` to `MediaType` (https://github.com/developmentseed/titiler/pull/258)
 
 ## 0.1.0 (2021-02-17)